Implement two stacks using one array
Logic is to fill stack 1 from left to right and stack 2 from right to left in the array to utilize maximum space. There should always be at least on gap between top1 and top 2 so that an element can be inserted thus push operation has a condition if(top2-top>1).